Moved trk io in dipy.io
